First Code commit.
As seen in my code, today I began working on my assessment project, setting up the basic HTML and CSS files that I will need for the first phase of development. The html files I added were the major files that I would need for my navigation bar, which is what I will begin creating first. By referring to my specifications page from our previous assessment, I knew exactly which files to add, starting with Home.html as the base file, before adding Recently Viewed, Favorites and Log-In. Because the design has a text logo without a designated image, I simply added text to the navigation bar, before making this text a link back to the home page, which is a consistency with industry standards, as most modern websites make it easy to get back to the home page by clicking the logo. Because I wanted to stay true to the design process that I used during the first assessment, I chose to also add the toggle feature that I included in my mock-up design, which switches between golf courses and accommodation viewing for the website.