Second Phase of dummy testing website - Second Style of Toggle
In this updated code is my second attempt at creating a functioning toggle button that fits within the navigation bar of my main project. Similarly to my first attempt at creating this toggle button, the code worked fine when styled to fit the middle of the screen, but after adjusting the sizing the button broke, with the elements not remaining inside their designated container.